+++ code/portfolio/src/components/EnergyDrink/EnergyDrink.scss
@@ -1,10 +1,10 @@
 .EnergyDrink {
-  
+    margin:auto 10px;
     z-index: 2;
 
     img {
-        width: 250px;
-        height: 600px;
+        width: 220px;
+        height: 520px;
         transform: rotate(-10deg);
     }
 }
