+++ code/portfolio/src/components/Title/Title.scss
@@ -1,5 +1,5 @@
 .Title {
-    margin: 200px 250px 150px 200px;
+    margin: auto 250px;
     font-size: 60px;
     font-weight: 800;
     font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
