+++ code/portfolio/src/containers/Cards/Card.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Card = props => {
+    return (
+        <div className="Card" style={{ backgroundColor: props.color }}>
+            {props.children}
+        </div>
+    );
+}
+
+export default Card;
