+++ code/portfolio/src/containers/Cards/Card.scss
@@ -1,0 +1,8 @@
+.Card {
+    width: 400px;
+    height: 350px;
+    box-shadow: 2px 2px 2px #000;
+    background-color: transparent;
+    border: 1px solid #f1f1f1;
+    perspective: 1000px; /* Remove this if you don't want the 3D effect */
+}
