+++ code/portfolio/src/containers/Section/Section.scss
@@ -1,5 +1,4 @@
 .Section {
-    padding:30px;
     height:600px;
     width:100%;
     display:flex;
